Enable paintings to exist in world space
When painting is turned on, a brush mesh is instantiated and starts painting.
When painting is turned off, the brush mesh is unparented from the paint target, and is given the tag 'Mesh'. The user can then view the mesh from any angle. And start painting again with a new mesh.